safari-600.7-branch/LayoutTests/fast/dom/shadow/content-reprojection-fallback-crash-expected.txt

A crash should not happen when fallback elements are reprojected.

PASS distributedNodes.item(0) is shadowRoot1.getElementById("fallback")
PASS distributedNodes.item(1) is shadowRoot1.getElementById("div")
PASS successfullyParsed is true

TEST COMPLETE
